Extract method reference resolver
Move assembly-scoped MemberRef and MethodSpec identities, resolution caches, and shared metadata work budgets out of LibraryBodyAnalysisBuilder while preserving lifted-owner and per-method consumers.
