+++ src/ILInspector.Analysis.Tests/LibraryBodyIndexTests.cs
@@ -5129,7 +5129,7 @@ public class LibraryBodyIndexTests
 
         Assert.Equal(local, external);
         Assert.False(
-            LibraryBodyAnalysisBuilder
+            LibraryBodyMethodReferenceResolver
                 .SameMethodReferenceDeclaringType(
                     local,
                     external));
